log information request details
